Create data for calibration curves using a binary logistic regression framework with inverse probability of censoring weights

Creates the underlying data for the calibration plots. Observed event probabilities at time t.eval are estimated for inputted predicted transition probabilities tp.pred out of state j at time s. calc_calib_blr estimates calibration curves using a binary logistic framework in combination with landmarking and inverse probability of censoring weights. A choice between restricted cubic splines and loess smoothers for estimating the calibration curve can be made using curve.type.

Two datasets for the same cohort of inidividuals must be provided. A msdata format dataset generated using the mstate package. A data.frame with one row per individual, relevant variables for estimating the weights, and a time until censoring varaible (dtcens) and indicator (dtcens.s). Weights are estimated using a cox-proportional hazard model and assuming linear functional form of the variables defined in w.covs. We urge users to specify their own model for estimating the weights. Confidence intervals for the calibration curves can be estimated using bootstrapping. This procedure uses the internal method for estimating weights, we therefore encourage users to specify their own bootstrapping procedure, which incorporates their own model for estimating the weights. Details on how to do this are provided in the vignette.

Usage

calc_calib_blr(
  data.mstate,
  data.raw,
  j,
  s,
  t.eval,
  tp.pred,
  curve.type = "rcs",
  rcs.nk = 3,
  loess.span,
  loess.degree,
  weights = NULL,
  w.covs,
  w.landmark.type = "state",
  w.max = 10,
  w.stabilised = FALSE,
  w.max.follow = NULL,
  CI = FALSE,
  CI.R.boot,
  data.pred.plot = NULL,
  transitions.out = NULL
)

Arguments

data.mstate

Validation data in msdata format.

data.raw

Validation data in data.frame (one row per individual).

j

Landmark state at which predictions were made

s

Landmark time at which predictions were made

t.eval

Follow up time at which calibration is to be assessed

tp.pred

Matrix of predicted transition probabilities at time t.eval, if in state j at time s. There must be a seperate column for the predicted transition probabilities into every state, even if these predicted transition probabilities are 0.

curve.type

Whether calibration curves are estimated using restricted cubic splines ('rcs') or loess smoothers ('loess')

rcs.nk

Number of knots when curves are estimated using restricted cubic splines

loess.span

Span when curves are estimated using loess smoothers

loess.degree

Degree when curves are estimated using loess smoothers

weights

Vector of inverse probability of censoring weights

w.covs

Character vector of variable names to adjust for when calculating inverse probability of censoring weights

w.landmark.type

Whether weights are estimated in all individuals uncensored at time s ('all') or only in individuals uncensored and in state j at time s ('state')

w.max

Maximum bound for inverse probability of censoring weights

w.stabilised

Indicates whether inverse probability of censoring weights should be stabilised or not

w.max.follow

Maximum follow up for model calculating inverse probability of censoring weights. Reducing this to t.eval + 1 may aid in the proportional hazards assumption being met in this model.

CI

Size of confidence intervals as a %

CI.R.boot

Number of bootstrap replicates when estimating the confidence interval for the calibration curve

data.pred.plot

Dataframe or matrix of predicted risks for each possible transition over which to plot the calibration curves. Must have one column for every possible transition.

transitions.out

Transitions for which to calculate calibration curves. Will do all possible transitions if left as NULL.