Calculate inverse probability of censoring weights at time t.eval for a landmark group of individuals.

Primarily used internally, this function has been exported to allow users to reproduce results in the vignette when estimating confidence intervals using bootstrapping manually.

Fits a cox proportional hazards model to individuals in a landmark cohort, predicting the probability of being censored at time t.eval. This landmark cohort may either be all individuals uncensored at time s, or those uncensored and in state j at time s. All predictors in w.covs are assumed to have a linear effect on the hazard. Weights are estimated for all individuals in data.raw, even if they will not be used in the analysis as they do not meet the landmarking requirements. If an individual enters an absorbing state prior to t.eval, we estimate the probability of being censored before the time of entry into the absorbing state, rather than at t.eval. Details on this are provided in the associated vignette.

Usage

calc_weights(
  data.mstate,
  data.raw,
  covs,
  t.eval,
  s,
  landmark.type = NULL,
  j = NULL,
  max.weight = 10,
  stabilised = FALSE,
  max.follow = NULL
)

Arguments

data.mstate

Validation data in msdata format

data.raw

Validation data in data.frame (one row per individual)

covs

Character vector of variable names to adjust for when calculating inverse probability of censoring weights

t.eval

Follow up time at which to calculate weights

s

Landmark time at which predictions were made

landmark.type

Whether weights are estimated in all individuals uncensored at time s ('all') or only in individuals uncensored and in state j at time s ('state')

j

Landmark state at which predictions were made (only required in landmark.type = 'state')

max.weight

Maximum bound for weights

stabilised

Indicates whether weights should be stabilised or not

max.follow

Maximum follow up for model calculating inverse probability of censoring weights. Reducing this to t.eval + 1 may aid in the proportional hazards assumption being met in this model.