Estimate calibration curves for a multistate model using binary logistic regression calibration techniques and inverse probability of censoring weights.'

Creates the underlying data for the calibration curves. calib_blr estimates the observed event probabilities for a given set of predicted transition probabilities in a cohort of interest. This is done using techniques for assessing calibration of binary logistic regression models, in combination with inverse probability of censoring weights and landmarking.

Usage

calib_blr_SAVE(
  data.mstate,
  data.raw,
  j,
  s,
  t,
  tp.pred,
  curve.type = "rcs",
  rcs.nk = 3,
  loess.span = 0.75,
  loess.degree = 2,
  weights = NULL,
  w.function = NULL,
  w.covs = NULL,
  w.landmark.type = "state",
  w.max = 10,
  w.stabilised = FALSE,
  w.max.follow = NULL,
  CI = FALSE,
  CI.type = "bootstrap",
  CI.R.boot = NULL,
  data.pred.plot = NULL,
  transitions.out = NULL,
  ...
)

Arguments

data.mstate

Validation data in msdata format

data.raw

Validation data in data.frame (one row per individual)

j

Landmark state at which predictions were made

s

Landmark time at which predictions were made

t

Follow up time at which calibration is to be assessed

tp.pred

Matrix of predicted transition probabilities at time t, if in state j at time s. There must be a seperate column for the predicted transition probabilities into every state, even if these predicted transition probabilities are 0.

curve.type

Whether calibration curves are estimated using restricted cubic splines ('rcs') or loess smoothers ('loess')

rcs.nk

Number of knots when curves are estimated using restricted cubic splines

loess.span

Span when curves are estimated using loess smoothers

loess.degree

Degree when curves are estimated. using loess smoothers

weights

Vector of inverse probability of censoring weights

w.function

Custom function for estimating the inverse probability of censoring weights

w.covs

Character vector of variable names to adjust for when calculating inverse probability of censoring weights

w.landmark.type

Whether weights are estimated in all individuals uncensored at time s ('all') or only in individuals uncensored and in state j at time s ('state')

w.max

Maximum bound for inverse probability of censoring weights

w.stabilised

Indicates whether inverse probability of censoring weights should be stabilised or not

w.max.follow

Maximum follow up for model calculating inverse probability of censoring weights. Reducing this to t + 1 may aid in the proportional hazards assumption being met in this model.

CI

Size of confidence intervals as a %

CI.type

Method for estimating confidence interval (currently restricted to bootstrap)

CI.R.boot

Number of bootstrap replicates when estimating the confidence interval for the calibration curve

data.pred.plot

Data frame or matrix of predicted risks for each possible transition over which to plot the calibration curves. Must have one column for every possible transition.

transitions.out

Transitions for which to calculate calibration curves. Will do all possible transitions if left as NULL.

...

Extra arguments to be passed to w.function (custom function for estimating weights)

Details

Observed event probabilities at time t are estimated for predicted transition probabilities tp.pred out of state j at time s. calib_blr estimates calibration curves using techniques for assessing the calibration of binary logistic regression models (Van Calster et al., 2016). A choice between restricted cubic splines and loess smoothers for estimating the calibration curve can be made using curve.type. Landmarking is applied to only assess calibration in individuals who are uncensored and in state j at time s. Censoring is dealt with using inverse probability of censoring weights (Hernan M, Robins J, 2020).

Two datasets for the same cohort of inidividuals must be provided. Firstly data.mstate must be a dataset of class msdata, generated using the [mstate] package. This dataset is used to apply the landmarking. Secondly, data.raw must be a data.frame with one row per individual, containing the desired variables for estimating the weights, and variables for the time until censoring (dtcens), and an indicator for censoring dtcens.s, where (dtcens.s = 1) if