Estimate calibration curves for a multistate model using pseudo-values.

Creates the underlying data for the calibration curves. calib_pv estimates the observed event probabilities for a given set of predicted transition probabilities in a cohort of interest. This is done using techniques for assessing calibration of binary logistic regression models, in combination with inverse probability of censoring weights and landmarking.

Usage

calib_pv(
  data.mstate,
  data.raw,
  j,
  s,
  t.eval,
  tp.pred,
  curve.type = "rcs",
  rcs.nk = 3,
  loess.span = 0.75,
  loess.degree = 2,
  group.vars = NULL,
  n.pctls = NULL,
  CI = FALSE,
  CI.type = NULL,
  CI.R.boot = NULL,
  data.pred.plot = NULL,
  transitions.out = NULL
)

Arguments

data.mstate

Validation data in msdata format

data.raw

Validation data in data.frame (one row per individual)

j

Landmark state at which predictions were made

s

Landmark time at which predictions were made

t.eval

Follow up time at which calibration is to be assessed

tp.pred

Matrix of predicted transition probabilities at time t.eval, if in state j at time s. There must be a seperate column for the predicted transition probabilities into every state, even if these predicted transition probabilities are 0.

curve.type

Whether calibration curves are estimated using restricted cubic splines ('rcs') or loess smoothers ('loess')

rcs.nk

Number of knots when curves are estimated using restricted cubic splines

loess.span

Span when curves are estimated using loess smoothers

loess.degree

Degree when curves are estimated. using loess smoothers

group.vars

Baseline variables to define groups within which to estimate pseudo-values

n.pctls

Number of percentiles to group individuals by with respect to predicted transition probabilities when estimating pseudo-values

CI

Size of confidence intervals as a %

CI.type

Way confidence interval is calculated (bootstrap or parametric)

CI.R.boot

Number of bootstrap replicates when estimating the confidence interval for the calibration curve using bootstrapping

data.pred.plot

Data frame or matrix of predicted risks for each possible transition over which to plot the calibration curves. Must have one column for every possible transition.

transitions.out

Transitions for which to calculate calibration curves. Will do all possible transitions if left as NULL.

Value

calib_pv returns a list containing two elements: plotdata and metadata. The plotdata element contains the data for the calibration curves. This will itself be a list with each element containing calibration plot data for the transition probabilities into each of the possible states. Each list element contains patient ids (id) from data.raw, the predicted transition probabilities (pred) and the estimated observed event probabilities (obs). If a confidence interval is requested, upper (obs.upper) and lower (obs.lower) bounds for the observed event probabilities are also returned. If data.pred.plot is defined manually, column (id) is not returned. The metadata element contains metadata including: a vector of the possible transitions, a vector of which transitions calibration curves have been estimated for, the size of the confidence interval, the method for estimating the calibration curve and other user specified information.

Details

Observed event probabilities at time t.eval are estimated for predicted transition probabilities tp.pred out of state j at time s. calib_pv estimates the observed event probabilities using pseudo-values. Calibraiton curves are generatd by regression the pseudo-values for the transition probabilities on the predicted transition probabilities. REF XXXX. Currently calibration curves can only be produced using loess smoothers. This will be updated to include restricted cubic splines. XXXX Landmarking is applied to only assess calibration in individuals who are uncensored and in state j at time s.

Two datasets for the same cohort of inidividuals must be provided. Firstly data.mstate must be a dataset of class msdata, generated using the [mstate] package. This dataset is used to apply the landmarking. Secondly, data.raw must be a data.frame with one row per individual, containing the desired variables for calculating pseudo-values within (no baseline variables required if group.vars = NULL). Confidence intervals for the calibration curves can be estimated using bootstrapping.

The calibration curves can be plotted using plot.calib_pv.

Examples

# Estimate calibration curves for the predicted transition
# probabilities at time t.eval = 1826, when predictions were made at time
# s = 100 in state j = 3. These predicted transition probabilities are stored in tps100.

# Extract the predicted transition probabilities out of state j = 1
tp.pred <- dplyr::select(dplyr::filter(tps100, j == 3), any_of(paste("pstate", 1:6, sep = "")))

# Now estimate the observed event probabilities for each possible transition.

dat.calib.pseudo <- calib_pv(data.mstate = msebmtcal,
  data.raw = ebmtcal,
  j = 3,
  s = 100,
  t.eval = 1826,
  tp.pred = tp.pred,
  group.vars = c("year"),
  n.pctls = 2,
  loess.span = 0.75, loess.degree = 2)

# The data for each calibration curve are stored in the "plotdata" list
# element.
str(dat.calib.pseudo)
#> List of 2
#>  $ plotdata:List of 4
#>   ..$ state3:'data.frame':	413 obs. of  3 variables:
#>   .. ..$ id  : num [1:413] 3 13 18 22 23 24 25 26 27 28 ...
#>   .. ..$ pred: num [1:413] 0.537 0.537 0.394 0.43 0.43 ...
#>   .. ..$ obs : num [1:413] 0.529 0.529 0.422 0.445 0.445 ...
#>   ..$ state4:'data.frame':	413 obs. of  3 variables:
#>   .. ..$ id  : num [1:413] 3 13 18 22 23 24 25 26 27 28 ...
#>   .. ..$ pred: num [1:413] 0.1332 0.1334 0.0477 0.0516 0.0516 ...
#>   .. ..$ obs : num [1:413] 0.1353 0.1355 0.0521 0.0545 0.0545 ...
#>   ..$ state5:'data.frame':	413 obs. of  3 variables:
#>   .. ..$ id  : num [1:413] 3 13 18 22 23 24 25 26 27 28 ...
#>   .. ..$ pred: num [1:413] 0.105 0.105 0.133 0.172 0.172 ...
#>   .. ..$ obs : num [1:413] 0.134 0.1342 0.1455 0.0879 0.0878 ...
#>   ..$ state6:'data.frame':	413 obs. of  3 variables:
#>   .. ..$ id  : num [1:413] 3 13 18 22 23 24 25 26 27 28 ...
#>   .. ..$ pred: num [1:413] 0.225 0.225 0.426 0.347 0.347 ...
#>   .. ..$ obs : num [1:413] 0.196 0.196 0.585 0.404 0.404 ...
#>  $ metadata:List of 11
#>   ..$ valid.transitions   : num [1:4] 3 4 5 6
#>   ..$ assessed.transitions: num [1:4] 3 4 5 6
#>   ..$ curve.type          : chr "rcs"
#>   ..$ CI                  : logi FALSE
#>   ..$ CI.type             : NULL
#>   ..$ CI.R.boot           : NULL
#>   ..$ j                   : num 3
#>   ..$ s                   : num 100
#>   ..$ t.eval              : num 1826
#>   ..$ group.vars          : chr "year"
#>   ..$ n.pctls             : num 2
#>  - attr(*, "class")= chr "calib_pv"

#'