Create data for calibration curves using a multinomial logistic regression framework with inverse probability of censoring weights

Creates the underlying data for the calibration plots. Observed event probabilities at time t are estimated for inputted predicted transition probabilities tp.pred out of state j at time s. calib_mlr estimates calibration scatter plots using a multinomial logistic framework in combination with landmarking and inverse probability of censoring weights.

Two datasets for the same cohort of inidividuals must be provided. A msdata format dataset generated using the mstate package. A data.frame with one row per individual, relevant variables for estimating the weights, and a time until censoring varaible (dtcens) and indicator (dtcens.s). Weights are estimated using a cox-proportional hazard model and assuming linear functional form of the variables defined in w.covs. We urge users to specify their own modwl for estimating the weights. Confidence intervals for the calibration scatter plots cannot be produced as it is currently unclear how to present such data.

Usage

calib_mlr(
  data.mstate,
  data.raw,
  j,
  s,
  t,
  tp.pred,
  smoother.type = "sm.ps",
  ps.int = 4,
  degree = 3,
  s.df = 4,
  niknots = 4,
  weights = NULL,
  w.function = NULL,
  w.covs = NULL,
  w.landmark.type = "state",
  w.max = 10,
  w.stabilised = FALSE,
  w.max.follow = NULL,
  ...
)

Arguments

data.mstate

Validation data in msdata format

data.raw

Validation data in data.frame (one row per individual)

j

Landmark state at which predictions were made

s

Landmark time at which predictions were made

t

Follow up time at which calibration is to be assessed

tp.pred

Vector of predicted transition probabilities at time t

smoother.type

Type of smoothing applied. Takes values s (see s), sm.ps (see sm.ps) or sm.os (see sm.os).

ps.int

the number of equally-spaced B spline intervals in the vector spline smoother (see sm.ps)

degree

the degree of B-spline basis in the vector spline smoother (see sm.ps)

s.df

degrees of freedom of vector spline (see s)

niknots

number of interior knots (see sm.os)

weights

Vector of inverse probability of censoring weights

w.function

Custom function for estimating the inverse probability of censoring weights

w.covs

Character vector of variable names to adjust for when calculating inverse probability of censoring weights

w.landmark.type

Whether weights are estimated in all individuals uncensored at time s ('all') or only in individuals uncensored and in state j at time s ('state')

w.max

Maximum bound for inverse probability of censoring weights

w.stabilised

Indicates whether inverse probability of censoring weights should be stabilised or not

w.max.follow

Maximum follow up for model calculating inverse probability of censoring weights. Reducing this to t + 1 may aid in the proportional hazards assumption being met in this model.

...

Extra arguments to be passed to w.function (custom function for estimating weights)

Details

Observed event probabilities at time t are estimated for predicted transition probabilities tp.pred out of state j at time s. calib_mlr estimates calibration scatter plots uses a technique for assessing the calibration of multinomial logistic regression models, namely the nominal calibration framework of #' van Hoorde et al. (2014, 2015). Landmarking is applied to only assess calibration in individuals who are uncensored and in state j at time s. Censoring is dealt with using inverse probability of censoring weights.

Two datasets for the same cohort of inidividuals must be provided. Firstly data.mstate must be a dataset of class msdata, generated using the [mstate] package. This dataset is used to apply the landmarking. Secondly, data.raw must be a data.frame with one row per individual, containing the desired variables for estimating the weights, and variables for the time until censoring (dtcens), and an indicator for censoring dtcens.s, where (dtcens.s = 1) if an individual is censored at time dtcens, and dtcens.s = 0 otherwise. When an individual enters an absorbing state, this prevents censoring from happening (i.e. dtcens.s = 0). Unless the user specifies the weights using weights, the weights are estimated using a cox-proportional hazard model, assuming a linear functional form of the variables defined in w.covs. We urge users to specify their own model for estimating the weights. The weights argument must be a vector with length equal to the number of rows of data.raw.

Confidence intervals cannot be generated for the calibration scatter plots. While confidence intervals could be estimated for each data point, it is not clear how these would be plotted cohesively.

Calibration plots cannot be produced for specific transitions (i.e. transitions.out in calib_blr) because the nominal calibration framework (van Hoorde et al., 2014, 2015) assesses the calibration of all states simultaneously.

The calibration scatter plots can be plotted using plot.calib_mlr.

#' @returns calib_mlr returns a list containing two elements: plotdata and metadata. The plotdata element contains the data for the calibration scatter plots This will itself be a list with each element containing the data for the transition probabilities into each of the possible states. Each list element contains patient ids (id), the predicted transition probabilities (pred) and the estimated observed event probabilities (obs). The metadata element contains metadata including a vector of the possible transitions and other user specified information.

Examples

# Estimate calibration scatter plots for the predicted transition
# probabilities at time t = 1826, when predictions were made at time
# s = 0 in state j = 1. These predicted transition probabilities are stored in tps0.

# Extract the predicted transition probabilities out of state j = 1
tp.pred <- dplyr::select(dplyr::filter(tps0, j == 1), any_of(paste("pstate", 1:6, sep = "")))

# Now estimate the observed event probabilities for each possible transition.

dat.calib.mlr <-
calib_mlr(data.mstate = msebmtcal,
 data.raw = ebmtcal,
 j=1,
 s=0,
 t = 1826,
 tp.pred = tp.pred,
 w.covs = c("year", "agecl", "proph", "match"))

# The data for each calibration scatter plots are stored in the "plotdata"
# list element.
str(dat.calib.mlr)
#> List of 2
#>  $ plotdata:List of 6
#>   ..$ state1:'data.frame':	1778 obs. of  3 variables:
#>   .. ..$ id  : int [1:1778] 2 4 5 7 10 13 14 16 18 19 ...
#>   .. ..$ pred: num [1:1778] 0.114 0.1384 0.1233 0.0974 0.1137 ...
#>   .. ..$ obs : num [1:1778] 0.092 0.0977 0.0506 0.1051 0.0947 ...
#>   ..$ state2:'data.frame':	1778 obs. of  3 variables:
#>   .. ..$ id  : int [1:1778] 2 4 5 7 10 13 14 16 18 19 ...
#>   .. ..$ pred: num [1:1778] 0.232 0.184 0.161 0.212 0.232 ...
#>   .. ..$ obs : num [1:1778] 0.0963 0.1609 0.0812 0.1076 0.0981 ...
#>   ..$ state3:'data.frame':	1778 obs. of  3 variables:
#>   .. ..$ id  : int [1:1778] 2 4 5 7 10 13 14 16 18 19 ...
#>   .. ..$ pred: num [1:1778] 0.0844 0.0758 0.0551 0.0615 0.0844 ...
#>   .. ..$ obs : num [1:1778] 0.1789 0.1387 0.0772 0.1472 0.1811 ...
#>   ..$ state4:'data.frame':	1778 obs. of  3 variables:
#>   .. ..$ id  : int [1:1778] 2 4 5 7 10 13 14 16 18 19 ...
#>   .. ..$ pred: num [1:1778] 0.233 0.218 0.183 0.221 0.233 ...
#>   .. ..$ obs : num [1:1778] 0.217 0.171 0.351 0.145 0.211 ...
#>   ..$ state5:'data.frame':	1778 obs. of  3 variables:
#>   .. ..$ id  : int [1:1778] 2 4 5 7 10 13 14 16 18 19 ...
#>   .. ..$ pred: num [1:1778] 0.148 0.154 0.143 0.144 0.149 ...
#>   .. ..$ obs : num [1:1778] 0.187 0.202 0.23 0.15 0.182 ...
#>   ..$ state6:'data.frame':	1778 obs. of  3 variables:
#>   .. ..$ id  : int [1:1778] 2 4 5 7 10 13 14 16 18 19 ...
#>   .. ..$ pred: num [1:1778] 0.189 0.23 0.335 0.264 0.189 ...
#>   .. ..$ obs : num [1:1778] 0.229 0.23 0.21 0.345 0.233 ...
#>  $ metadata:List of 4
#>   ..$ valid.transitions: Named int [1:6] 1 2 3 4 5 6
#>   .. ..- attr(*, "names")= chr [1:6] "tp.pred1" "tp.pred2" "tp.pred3" "tp.pred4" ...
#>   ..$ j                : num 1
#>   ..$ s                : num 0
#>   ..$ t                : num 1826
#>  - attr(*, "class")= chr "calib_mlr"